refactor to make it easier to upload/test and change how to handle newman path
