+++ .gitignore
@@ -26,6 +26,7 @@ bld/
 # Visual Studio 2015 cache/options directory
 .vs/
 
+*/dist/
 node_modules
 .vscode
 *.vsix
