+++ NewmanPostman/package.json
@@ -2,11 +2,19 @@
   "name": "task",
   "version": "1.0.0",
   "description": "package.json for task lib",
+  "main": "dist/newmantask.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
+  "files": [
+    "src/",
+    "dist/",
+    "config/"
+  ],
   "dependencies": {
-    "vsts-task-lib": "^2.0.5"
+    "vsts-task-lib": "^2.0.5",
+    "reflect-metadata": "^0.1.12",
+    "tslib": "^1.9.3"
   },
   "devDependencies": {
     "@types/node": "^8.0.7",
