--- package.json
@@ -1,21 +1,0 @@
-{
-  "scripts": {
-    "package": "npm run clean  && npm run compile  && tfx extension create  --rev-version",
-    "gallery-publish": "tfx extension publish",
-    "gallery-publish-rev": "tfx extension publish --rev-version",
-    "clean": "rimraf ./*.vsix",
-    "compile": "tsc -p .",
-    "install-task-lib": "cd NewmanPostman && npm install --save-dev"
-  },
-  "devDependencies": {
-    "@types/node": "^8.0.7",
-    "tfx-cli": "^0.6.1",
-    "typescript": "2.3.4"
-  },
-  "name": "vstsexttask",
-  "private": true,
-  "version": "0.0.1",
-  "dependencies": {
-    "vsts-task-lib": "^2.1.0"
-  }
-}
