--- tsconfig.json
@@ -1,10 +1,0 @@
-{
-    "compilerOptions": {
-        "target": "ES6",
-        "module": "commonjs"
-    },
-    "exclude": [
-        "node_modules",
-        "NewmanPostman/node_modules"
-    ]
-}
